Add model relationships

+++ app/Http/Controllers/NoteController.php
@@ -16,7 +16,11 @@ class NoteController extends Controller
      */
     public function index()
     {
-        $notes = Note::where('user_id', Auth::id())->latest('updated_at')->paginate(3);
+        // $notes = Note::where('user_id', Auth::id())->latest('updated_at')->paginate(5);
+        // $notes = Auth::user()->notes()->latest('updated_at')->paginate(5);
+        /* Laravel 8 way */
+        $notes = Note::whereBelongsTo(Auth::user())->latest('updated_at')->paginate(5);
+
         return view('notes.index')->with('notes', $notes);
     }
 
@@ -60,6 +64,9 @@ class NoteController extends Controller
      */
     public function show(Note $note)
     {
+        if (!$note->user()->is(Auth::user())) {
+            return abort(403);
+        }
         return view('notes.show')->with('note', $note);
     }
 
@@ -71,6 +78,10 @@ class NoteController extends Controller
      */
     public function edit(Note $note)
     {
+        if (!$note->user()->is(Auth::user())) {
+            return abort(403);
+        }
+
         return view('notes.edit')->with('note', $note);
     }
 
@@ -83,6 +94,10 @@ class NoteController extends Controller
      */
     public function update(Request $request, Note $note)
     {
+        if (!$note->user()->is(Auth::user())) {
+            return abort(403);
+        }
+
         $request->validate([
             'title' => 'required|max:120',
             'text' => 'required'
@@ -104,6 +119,10 @@ class NoteController extends Controller
      */
     public function destroy(Note $note)
     {
+        if (!$note->user()->is(Auth::user())) {
+            return abort(403);
+        }
+
         $note->delete();
         return to_route('notes.index', $note)->with('success','Note deleted successfully.');
     }
